Add files via upload
ABOUT CONTACT MANAGEMENT SYSTEM::
This system is designed to store and manage contact information, including names, phone numbers, email addresses, and additional details. It provides a user-friendly interface for adding, editing, and viewing contacts.
The primary purpose of the Contact Management System is to streamline the storage, retrieval, and manipulation of contact data. By providing functionalities for insertion, deletion, and display of contacts, it aims to enhance contact organization and accessibility.